Move BlockIdSerializer to separate file
