+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Block.kt
@@ -1,13 +1,7 @@
 package com.swmansion.starknet.data.types
 
-import kotlinx.serialization.KSerializer
+import com.swmansion.starknet.data.serializers.BlockIdSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-import kotlinx.serialization.json.*
 
 @Serializable
 enum class BlockTag(val tag: String) {
@@ -37,28 +31,3 @@ sealed class BlockId() {
     }
 }
 
-internal class BlockIdSerializer() : KSerializer<BlockId> {
-    override fun deserialize(decoder: Decoder): BlockId {
-        val value = decoder.decodeString()
-
-        if (BlockTag.values().map { it.tag }.contains(value)) {
-            val tag = BlockTag.valueOf(value)
-            return BlockId.Tag(tag)
-        }
-
-        return BlockId.Hash(Felt.fromHex(value))
-    }
-
-    override val descriptor: SerialDescriptor
-        get() = PrimitiveSerialDescriptor("BlockHashOrTag", PrimitiveKind.STRING)
-
-    override fun serialize(encoder: Encoder, value: BlockId) {
-        require(encoder is JsonEncoder)
-        val element = when (value) {
-            is BlockId.Tag -> encoder.json.encodeToJsonElement(value.toString())
-            is BlockId.Hash -> buildJsonObject { put("block_hash", value.toString()) }
-            is BlockId.Number -> buildJsonObject { put("block_number", value.blockNumber) }
-        }
-        encoder.encodeJsonElement(element)
-    }
-}
